fix: 즐겨찾기 조회시 발생하는 LazyInitializationException 처리 Moive의 averageRating Type float -> double 변경, Series와 통일함

+++ backend/src/main/java/com/example/Flicktionary/domain/movie/dto/MovieResponse.java
@@ -11,7 +11,7 @@ public class MovieResponse {
     private final long tmdbId;
     private final String title;
     private final String posterPath;
-    private final float averageRating;
+    private final double averageRating;
     private final int ratingCount;
 
     public MovieResponse(Movie movie) {
